+++ internal/provider/function_sort_test.go
@@ -5,6 +5,7 @@
 package provider
 
 import (
+	"regexp"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
@@ -21,14 +22,12 @@ func TestSemversSortFunction_Known(t *testing.T) {
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: `
-        locals {
+				Config: `locals {
           list = ["0.1.1-rc1+a231f59", "0.1.1", "0.1.10", "0.1.2-rc1"]
         }
         output "semvers_sorted" {
           value = provider::semvers::sort(local.list)
-        }
-        `,
+        }`,
 				ConfigStateChecks: []statecheck.StateCheck{
 					statecheck.ExpectKnownOutputValue(
 						"semvers_sorted",
@@ -42,14 +41,12 @@ func TestSemversSortFunction_Known(t *testing.T) {
 				},
 			},
 			{
-				Config: `
-        locals {
+				Config: `locals {
           list = ["2","2.0", "2.0.0", "v2", "v2.0", "v2.0.0"]
         }
         output "semvers_deduped" {
           value = provider::semvers::sort(local.list)
-        }
-        `,
+        }`,
 				ConfigStateChecks: []statecheck.StateCheck{
 					statecheck.ExpectKnownOutputValue(
 						"semvers_deduped",
@@ -58,14 +55,12 @@ func TestSemversSortFunction_Known(t *testing.T) {
 				},
 			},
 			{
-				Config: `
-        locals {
+				Config: `locals {
           list = ["2+abc", "2.0.0", "v2", "v2.0+abc"]
         }
         output "semvers_deduped" {
           value = provider::semvers::sort(local.list)
-        }
-        `,
+        }`,
 				ConfigStateChecks: []statecheck.StateCheck{
 					statecheck.ExpectKnownOutputValue(
 						"semvers_deduped",
@@ -77,47 +72,39 @@ func TestSemversSortFunction_Known(t *testing.T) {
 	})
 }
 
-// func TestSemversSortFunction_Null(t *testing.T) {
-//   resource.UnitTest(t, resource.TestCase{
-//     TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-//       tfversion.SkipBelow(tfversion.Version1_8_0),
-//     },
-//     ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-//     Steps: []resource.TestStep{
-//       {
-//         Config: `
-//         output "test" {
-//           value = provider::semvers::sort_semvers(null)
-//         }
-//         `,
-//         // The parameter does not enable AllowNullValue
-//         ExpectError: regexp.MustCompile(`argument must not be null`),
-//       },
-//     },
-//   })
-// }
+func TestSemversSortFunction_Invalid(t *testing.T) {
+	t.Parallel()
 
-// func TestSemversSortFunction_Unknown(t *testing.T) {
-//   resource.UnitTest(t, resource.TestCase{
-//     TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-//       tfversion.SkipBelow(tfversion.Version1_8_0),
-//     },
-//     ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-//     Steps: []resource.TestStep{
-//       {
-//         Config: `
-//         resource "terraform_data" "test" {
-//           input = "testvalue"
-//         }
-
-//         output "test" {
-//           value = provider::semvers::sort_semvers(terraform_data.test.output)
-//         }
-//         `,
-//         Check: resource.ComposeAggregateTestCheckFunc(
-//           resource.TestCheckOutput("test", "testvalue"),
-//         ),
-//       },
-//     },
-//   })
-// }
+	resource.UnitTest(t, resource.TestCase{
+		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
+			tfversion.SkipBelow(tfversion.Version1_8_0),
+		},
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: `output "results" {
+          value = provider::semvers::sort(["0.1.0", "blah"])
+        }`,
+				ExpectError: regexp.MustCompile(`Invalid Semantic Version`),
+			},
+			{
+				Config: `output "results" {
+          value = provider::semvers::sort(null)
+        }`,
+				ExpectError: regexp.MustCompile(`Invalid function argument`),
+			},
+			{
+				Config: `output "results" {
+          value = provider::semvers::sort(true)
+        }`,
+				ExpectError: regexp.MustCompile(`Invalid function argument`),
+			},
+			{
+				Config: `output "results" {
+          value = provider::semvers::sort({"one": 1})
+        }`,
+				ExpectError: regexp.MustCompile(`Invalid function argument`),
+			},
+		},
+	})
+}